refactor redirect helpers
