+++ packages/next/src/client/components/is-next-router-error.ts
@@ -2,7 +2,7 @@ import {
   isHTTPAccessFallbackError,
   type HTTPAccessFallbackError,
 } from './http-access-fallback/http-access-fallback'
-import { isRedirectError, type RedirectError } from './redirect'
+import { isRedirectError, type RedirectError } from './redirect-error'
 
 /**
  * Returns true if the error is a navigation signal error. These errors are
